+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Use "Conversation" visibility by default when replying in subscribers-only conversation.
+- Display error message on "Manage subscriptions" page when Monero address is not valid.
 
 ## [3.12.0] - 2024-12-19