+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Updated translations ([#61](https://codeberg.org/silverpill/mitra-web/pulls/61)).
+
 ## [3.13.0] - 2024-12-28
 
 ### Added