+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [3.13.1] - 2025-01-03
+
 ### Changed
 
 - Updated translations ([#61](https://codeberg.org/silverpill/mitra-web/pulls/61)).